+++ CHANGELOG.md
@@ -9,7 +9,7 @@ headers from the dumper output.
 - The dump property `response.status_text` was renamed to `response.statusText`
   to match the JavaScript `Response` constructor.
 
-- Add option to specify dump file(s) to run from CLI as an alternative to running all dumps.
+- Add option to specify dump files to run from CLI as an alternative to running all dumps.
 
 ## 5.3.0 (2023-03-07)